+++ darwin/Classes/UniversalBle.g.swift
@@ -259,6 +259,7 @@ protocol UniversalBlePlatformChannel {
   func pair(deviceId: String) throws
   func unPair(deviceId: String) throws
   func getConnectedDevices(withServices: [String], completion: @escaping (Result<[UniversalBleScanResult], Error>) -> Void)
+  func isConnected(deviceId: String) throws -> Bool
 }
 
 /// Generated setup class from Pigeon to handle messages through the `binaryMessenger`.
@@ -514,6 +515,21 @@ class UniversalBlePlatformChannelSetup {
       }
     } else {
       getConnectedDevicesChannel.setMessageHandler(nil)
+    }
+    let isConnectedChannel = FlutterBasicMessageChannel(name: "dev.flutter.pigeon.universal_ble.UniversalBlePlatformChannel.isConnected\(channelSuffix)", binaryMessenger: binaryMessenger, codec: codec)
+    if let api = api {
+      isConnectedChannel.setMessageHandler { message, reply in
+        let args = message as! [Any?]
+        let deviceIdArg = args[0] as! String
+        do {
+          let result = try api.isConnected(deviceId: deviceIdArg)
+          reply(wrapResult(result))
+        } catch {
+          reply(wrapError(error))
+        }
+      }
+    } else {
+      isConnectedChannel.setMessageHandler(nil)
     }
   }
 }