+++ windows/src/universal_ble_plugin.cpp
@@ -14,6 +14,7 @@
 #include "helper/utils.h"
 #include "helper/universal_enum.h"
 #include "generated/universal_ble.g.h"
+#include "pin_entry.h"
 
 namespace universal_ble
 {
@@ -162,6 +163,15 @@ namespace universal_ble
       return FlutterError("Bluetooth is not available");
     }
   };
+
+  ErrorOr<bool> UniversalBlePlugin::IsConnected(const std::string &device_id)
+  {
+    auto it = connectedDevices.find(_str_to_mac_address(device_id));
+    if (it == connectedDevices.end())
+      return false;
+    auto deviceAgent = *it->second;
+    return deviceAgent.device.ConnectionStatus() == BluetoothConnectionStatus::Connected;
+  }
 
   std::optional<FlutterError> UniversalBlePlugin::Connect(const std::string &device_id)
   {
@@ -408,37 +418,24 @@ namespace universal_ble
       auto device = async_get(BluetoothLEDevice::FromBluetoothAddressAsync(_str_to_mac_address(device_id)));
       std::cout << "PairLog: Device to be paired was found" << std::endl;
       auto deviceInformation = device.DeviceInformation();
-      bool isPaired = deviceInformation.Pairing().IsPaired();
-      if (isPaired)
+
+      if (deviceInformation.Pairing().IsPaired())
+      {
         return FlutterError("PairLog: Device is already paired");
-      bool canPair = deviceInformation.Pairing().CanPair();
-      if (!canPair)
+      }
+      if (!deviceInformation.Pairing().CanPair())
+      {
         return FlutterError("PairLog: Device is not pairable");
+      }
 
-      auto customPairing = deviceInformation.Pairing().Custom();
-      winrt::event_token token = customPairing.PairingRequested([this](const Enumeration::DeviceInformationCustomPairing &sender, const Enumeration::DevicePairingRequestedEventArgs &eventArgs)
-                                                                {
-                                                                  std::cout << "PairLog: Pairing requested" << std::endl;
-                                                                // eventArgs.AcceptWithPasswordCredential(nullptr, nullptr);
-                                                                // eventArgs.Pin();
-                                                                // Accept all pairing request
-                                                                eventArgs.Accept(); });
-      // DevicePairingKinds => None, ConfirmOnly, DisplayPin, ProvidePin, ConfirmPinMatch, ProvidePasswordCredential
-      // DevicePairingProtectionLevel =>  Default, None, Encryption, EncryptionAndAuthentication
-      std::cout << "PairLog: Trying to pair" << std::endl;
-      auto async_c = customPairing.PairAsync(
-          Enumeration::DevicePairingKinds::ConfirmOnly);
-      async_c.Completed([this, customPairing, token, device_id](IAsyncOperation<DevicePairingResult> const &sender, AsyncStatus const args)
-                        {
-                          auto result = sender.GetResults();
-                          std::cout << "PairLog: Received pairing status" << std::endl;
-                          customPairing.PairingRequested(token);
-                          auto isPaired = result.Status() == Enumeration::DevicePairingResultStatus::Paired;
-
-                          // Post to main thread
-                          std::string errorStr = parsePairingFailError(result);
-                          uiThreadHandler_.Post([device_id, isPaired, errorStr]
-                                                { callbackChannel->OnPairStateChange(device_id, isPaired, &errorStr, SuccessCallback, ErrorCallback); }); });
+      if (isWindows11OrGreater())
+      {
+        PairAsync(device_id, deviceInformation);
+      }
+      else
+      {
+        CustomPairAsync(device_id, deviceInformation);
+      }
       return std::nullopt;
     }
     catch (const FlutterError &err)
@@ -552,6 +549,63 @@ namespace universal_ble
     {
       std::cout << "Bluetooth is not available" << std::endl;
     }
+  }
+
+  winrt::fire_and_forget UniversalBlePlugin::PairAsync(std::string device_id, const DeviceInformation deviceInformation)
+  {
+    try
+    {
+      auto result = co_await deviceInformation.Pairing().PairAsync();
+      std::cout << "PairLog: Received pairing status" << std::endl;
+      auto isPaired = result.Status() == Enumeration::DevicePairingResultStatus::Paired;
+      std::string errorStr = parsePairingFailError(result);
+      uiThreadHandler_.Post([device_id, isPaired, errorStr]
+                            { callbackChannel->OnPairStateChange(device_id, isPaired, &errorStr, SuccessCallback, ErrorCallback); });
+    }
+    catch (...)
+    {
+      std::cout << "PairLog: Unknown error" << std::endl;
+    }
+  }
+
+  winrt::fire_and_forget UniversalBlePlugin::CustomPairAsync(std::string device_id, const DeviceInformation deviceInformation)
+  {
+    try
+    {
+      auto customPairing = deviceInformation.Pairing().Custom();
+      winrt::event_token token = customPairing.PairingRequested({this, &UniversalBlePlugin::PairingRequestedHandler});
+      std::cout << "PairLog: Trying to pair" << std::endl;
+      DevicePairingProtectionLevel protectionLevel = deviceInformation.Pairing().ProtectionLevel();
+      // DevicePairingKinds => None, ConfirmOnly, DisplayPin, ProvidePin, ConfirmPinMatch, ProvidePasswordCredential
+      auto result = co_await customPairing.PairAsync(DevicePairingKinds::ConfirmOnly | DevicePairingKinds::ProvidePin, protectionLevel);
+      std::cout << "PairLog: Got Pair Result" << std::endl;
+      DevicePairingResultStatus status = result.Status();
+      customPairing.PairingRequested(token);
+      auto isPaired = status == Enumeration::DevicePairingResultStatus::Paired;
+      std::string errorStr = parsePairingFailError(result);
+      uiThreadHandler_.Post([device_id, isPaired, errorStr]
+                            { callbackChannel->OnPairStateChange(device_id, isPaired, &errorStr, SuccessCallback, ErrorCallback); });
+    }
+    catch (...)
+    {
+      std::cout << "PairLog Error: Pairing Failed" << std::endl;
+    }
+  }
+
+  void UniversalBlePlugin::PairingRequestedHandler(DeviceInformationCustomPairing sender, DevicePairingRequestedEventArgs eventArgs)
+  {
+    std::cout << "PairLog: Got PairingRequest" << std::endl;
+    DevicePairingKinds kind = eventArgs.PairingKind();
+    if (kind != DevicePairingKinds::ProvidePin)
+    {
+      eventArgs.Accept();
+      return;
+    }
+
+    std::cout << "PairLog: Trying to get pin from user" << std::endl;
+    hstring pin = askForPairingPin();
+    std::wcout << "PairLog: Got Pin: " << pin.c_str() << std::endl;
+    eventArgs.Accept(pin);
   }
 
   // Send device to callback channel